# bravo-api-spec.yaml
swagger: '2.2.1'
info:
  title: BRAvo API
  description: API description in **Markdown**.
  version: 1.0.0
host: vm0131.france-bioinformatique.fr
schemes:
  - http
  - https
paths:
  /automatic/upstream:
    post:
      summary: Run upstream assembly
      description: Returns an upstream regulation graph based on a list of target genes.
      parameters:
        - in: formData
          name: genes
          type: array
          required: true
          items:
            type: string
            format: json
            uniqueItems: true
            enum:
              - '[MYC,GALT]'
              - '[ENSG00000158669,HGNC:4135,ENSG00000116133]'
          description: List of biological entities (target)
        - in: formData
          name: type
          type: string
          required: true
          enum:
            - id
            - name
          description: Define the type of entities
      produces:
        - application/json
      consumes:
        - application/x-www-form-urlencoded
      responses:
        '200':
          description: "OK"
        '500':
          description: "Error while processing automatic upstream network assembly"
  /automatic/downstream:
    post:
      summary: Run downstream assembly
      description: Returns a downstream regulation graph based on a list of source genes.
      parameters:
        - in: formData
          name: genes
          type: array
          required: true
          items:
            type: string
            format: json
            uniqueItems: true
            enum:
              - '[MYC,GALT]'
              - '[ENSG00000158669,HGNC:4135,ENSG00000116133]'
          description: List of biological entities (source)
        - in: formData
          name: type
          type: string
          required: true
          description: Define the type of entities
          enum:
            - id
            - name
      produces:
        - application/json
      consumes:
        - application/x-www-form-urlencoded
      responses:
        '200':
          description: "OK"
        '500':
          description: "Error while processing automatic upstream network assembly"